move findDOMNode util to correct file
