+++ src/Tablo/utils.ts
@@ -1,6 +1,3 @@
-import { ReactInstance } from "react";
-import * as ReactDOM from "react-dom";
-
 export const getArrayChildren: (
   maybeArrayChildren?:
     | React.ReactElement<any>
@@ -12,7 +9,3 @@ export const getArrayChildren: (
       ? maybeArrayChildren
       : [maybeArrayChildren]
     : undefined;
-
-export const findDOMNode = <E extends Element = Element>(
-  ref: ReactInstance
-): E => ReactDOM.findDOMNode(ref) as E;
